requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Caching>=2.0.0
Pillow>=9.0.0
requests>=2.25.0
python-dotenv>=0.19.0
Flask-WTF>=1.0.0
WTForms>=3.0.0
google-generativeai>=0.3.0
